fix: integrate title bar scroll into DockObserver
