+++ DockDoor/AppDelegate.swift
@@ -10,7 +10,6 @@ class AppDelegate: NSObject, NSApplicationDelegate {
     private var windowSeeder: WindowSeeder?
     private var previewCoordinator: SharedPreviewWindowCoordinator?
     private var keybindHelper: KeybindHelper?
-    private var titleBarScrollObserver: TitleBarScrollObserver?
     private var activeAppIndicator: ActiveAppIndicatorCoordinator?
     private var statusBarItem: NSStatusItem?
     private var updaterController: SPUStandardUpdaterController
@@ -57,8 +56,6 @@ class AppDelegate: NSObject, NSApplicationDelegate {
             removeMenuBar()
         }
 
-        titleBarScrollObserver = TitleBarScrollObserver()
-
         if !Defaults[.launched] {
             handleFirstTimeLaunch()
         } else {
@@ -68,7 +65,8 @@ class AppDelegate: NSObject, NSApplicationDelegate {
             let needsDockObserver = Defaults[.enableDockPreviews] ||
                 Defaults[.shouldHideOnDockItemClick] ||
                 Defaults[.enableCmdRightClickQuit] ||
-                Defaults[.enableDockScrollGesture]
+                Defaults[.enableDockScrollGesture] ||
+                Defaults[.enableTitleBarScrollGesture]
 
             if needsDockObserver {
                 let dockObs = DockObserver(previewCoordinator: currentPreviewCoordinator)
