+++ DockDoor/Utilities/DockObserver.swift
@@ -1,4 +1,5 @@
 import ApplicationServices
+import Carbon.HIToolbox.Events
 import Cocoa
 import Defaults
 
@@ -28,6 +29,19 @@ func handleSelectedDockItemChangedNotification(observer _: AXObserver, element _
 }
 
 final class DockObserver {
+    private enum VerticalScrollAction {
+        case maximize
+        case center
+    }
+
+    private struct PendingRestoreState {
+        let windowId: CGWindowID
+        let processIdentifier: pid_t
+        let action: VerticalScrollAction
+        let originalFrame: CGRect
+        let expirationDate: Date
+    }
+
     weak static var activeInstance: DockObserver?
     let previewCoordinator: SharedPreviewWindowCoordinator
 
@@ -53,8 +67,35 @@ final class DockObserver {
     var lastHoveredAppHadWindows: Bool = false
 
     // Scroll gesture state
-    private var lastScrollActionTime: Date = .distantPast
-    private let scrollActionDebounceInterval: TimeInterval = 0.3
+    private var lastDockScrollActionTime: Date = .distantPast
+    private let dockScrollActionDebounceInterval: TimeInterval = 0.3
+    private var lastTitleBarScrollActionTime: Date = .distantPast
+    private var pendingTitleBarRestoreState: PendingRestoreState?
+
+    private let titleBarScrollActionDebounceInterval: TimeInterval = 0.35
+    private let minimumTitleBarScrollDelta: Double = 0.15
+    private let fallbackTitleBarHeight: CGFloat = 48
+    private let maximumTitleBarHeight: CGFloat = 72
+
+    private var centeredWindowScale: CGFloat {
+        min(max(Defaults[.titleBarScrollCenteredWindowScale], 0.2), 1.0)
+    }
+
+    private var centeredWindowWidthScale: CGFloat {
+        min(max(Defaults[.titleBarScrollCenteredWindowWidthScale], 0.2), 1.0)
+    }
+
+    private var centeredWindowHeightScale: CGFloat {
+        min(max(Defaults[.titleBarScrollCenteredWindowHeightScale], 0.2), 1.0)
+    }
+
+    private var centeredWindowLockAspectRatio: Bool {
+        Defaults[.titleBarScrollCenteredWindowLockAspectRatio]
+    }
+
+    private var centeredWindowSizingMode: TitleBarCenteredWindowSizingMode {
+        Defaults[.titleBarScrollCenteredWindowSizingMode]
+    }
 
     private static func isSpecialControlsApp(_ bundleId: String?) -> Bool {
         bundleId == spotifyAppIdentifier ||
@@ -586,6 +627,10 @@ final class DockObserver {
         }
 
         if type == .scrollWheel {
+            if Defaults[.enableTitleBarScrollGesture], handleTitleBarScroll(event) {
+                return nil
+            }
+
             guard Defaults[.enableDockScrollGesture] else {
                 return Unmanaged.passUnretained(event)
             }
@@ -819,10 +864,10 @@ final class DockObserver {
         }
 
         let now = Date()
-        guard now.timeIntervalSince(lastScrollActionTime) >= scrollActionDebounceInterval else {
+        guard now.timeIntervalSince(lastDockScrollActionTime) >= dockScrollActionDebounceInterval else {
             return true // Still consume the event during debounce
         }
-        lastScrollActionTime = now
+        lastDockScrollActionTime = now
 
         if normalizedDeltaY > 0 {
             activateApp(app)
@@ -855,5 +900,154 @@ final class DockObserver {
             app.hide()
             previewCoordinator.hideWindow()
         }
+    }
+
+    private func handleTitleBarScroll(_ event: CGEvent) -> Bool {
+        let deltaX = event.getDoubleValueField(.scrollWheelEventDeltaAxis2)
+        let deltaY = event.getDoubleValueField(.scrollWheelEventDeltaAxis1)
+
+        guard abs(deltaX) > minimumTitleBarScrollDelta || abs(deltaY) > minimumTitleBarScrollDelta else {
+            return false
+        }
+
+        let mouseLocation = NSEvent.mouseLocation
+
+        guard let focusedWindow = WindowUtil.getFocusedWindowForFrontmostApp(),
+              let focusedWindowFrame = focusedWindow.currentWindowFrame(),
+              isPointInTitleBar(mouseLocation, window: focusedWindow, windowFrame: focusedWindowFrame)
+        else {
+            return false
+        }
+
+        let nsEvent = NSEvent(cgEvent: event)
+        let isNaturalScrolling = nsEvent?.isDirectionInvertedFromDevice ?? false
+        let normalizedDeltaX = isNaturalScrolling ? -deltaX : deltaX
+        let normalizedDeltaY = isNaturalScrolling ? -deltaY : deltaY
+
+        let now = Date()
+        guard now.timeIntervalSince(lastTitleBarScrollActionTime) >= titleBarScrollActionDebounceInterval else {
+            return true
+        }
+        lastTitleBarScrollActionTime = now
+
+        if abs(normalizedDeltaX) > abs(normalizedDeltaY) {
+            if normalizedDeltaX > 0 {
+                switchToNextSpace()
+            } else {
+                switchToPreviousSpace()
+            }
+            return true
+        }
+
+        if normalizedDeltaY > 0 {
+            handleTitleBarVerticalScroll(.maximize, for: focusedWindow, now: now)
+        } else {
+            handleTitleBarVerticalScroll(.center, for: focusedWindow, now: now)
+        }
+
+        return true
+    }
+
+    private func handleTitleBarVerticalScroll(_ action: VerticalScrollAction, for window: WindowInfo, now: Date) {
+        if let pendingTitleBarRestoreState,
+           pendingTitleBarRestoreState.windowId == window.id,
+           pendingTitleBarRestoreState.processIdentifier == window.app.processIdentifier,
+           pendingTitleBarRestoreState.action == action,
+           pendingTitleBarRestoreState.expirationDate > now
+        {
+            window.setWindowFrame(pendingTitleBarRestoreState.originalFrame)
+            self.pendingTitleBarRestoreState = nil
+            return
+        }
+
+        guard let originalFrame = window.currentWindowFrame() else {
+            performTitleBarVerticalScrollAction(action, on: window)
+            pendingTitleBarRestoreState = nil
+            return
+        }
+
+        performTitleBarVerticalScrollAction(action, on: window)
+        pendingTitleBarRestoreState = PendingRestoreState(
+            windowId: window.id,
+            processIdentifier: window.app.processIdentifier,
+            action: action,
+            originalFrame: originalFrame,
+            expirationDate: now.addingTimeInterval(Double(Defaults[.titleBarScrollRestoreWindowInterval]))
+        )
+    }
+
+    private func performTitleBarVerticalScrollAction(_ action: VerticalScrollAction, on window: WindowInfo) {
+        switch action {
+        case .maximize:
+            window.zoom()
+        case .center:
+            switch centeredWindowSizingMode {
+            case .uniform:
+                window.centerWindow(scale: centeredWindowScale)
+            case .separate:
+                window.centerWindow(
+                    widthScale: centeredWindowWidthScale,
+                    heightScale: centeredWindowHeightScale,
+                    lockAspectRatio: centeredWindowLockAspectRatio
+                )
+            }
+        }
+    }
+
+    private func isPointInTitleBar(_ point: CGPoint, window: WindowInfo, windowFrame: CGRect) -> Bool {
+        guard windowFrame.contains(point) else {
+            return false
+        }
+
+        let titleBarHeight = resolvedTitleBarHeight(for: window, windowFrame: windowFrame)
+        let titleBarFrame = CGRect(
+            x: windowFrame.minX,
+            y: windowFrame.maxY - titleBarHeight,
+            width: windowFrame.width,
+            height: titleBarHeight
+        )
+        return titleBarFrame.contains(point)
+    }
+
+    private func resolvedTitleBarHeight(for window: WindowInfo, windowFrame: CGRect) -> CGFloat {
+        let closeButton = window.closeButton ?? ((try? window.axElement.closeButton()) ?? nil)
+        guard let closeButton,
+              let closeButtonFrame = WindowInfo.currentWindowFrame(for: closeButton)
+        else {
+            return fallbackTitleBarHeight
+        }
+
+        let inferredHeight = windowFrame.maxY - closeButtonFrame.minY + 12
+        return min(max(inferredHeight, fallbackTitleBarHeight), maximumTitleBarHeight)
+    }
+
+    private func switchToPreviousSpace() {
+        postControlArrowKey(CGKeyCode(kVK_LeftArrow))
+    }
+
+    private func switchToNextSpace() {
+        postControlArrowKey(CGKeyCode(kVK_RightArrow))
+    }
+
+    private func postControlArrowKey(_ keyCode: CGKeyCode) {
+        guard let source = CGEventSource(stateID: .combinedSessionState) else {
+            return
+        }
+
+        let controlDown = CGEvent(keyboardEventSource: source, virtualKey: CGKeyCode(kVK_Control), keyDown: true)
+        controlDown?.flags = .maskControl
+
+        let arrowDown = CGEvent(keyboardEventSource: source, virtualKey: keyCode, keyDown: true)
+        arrowDown?.flags = .maskControl
+
+        let arrowUp = CGEvent(keyboardEventSource: source, virtualKey: keyCode, keyDown: false)
+        arrowUp?.flags = .maskControl
+
+        let controlUp = CGEvent(keyboardEventSource: source, virtualKey: CGKeyCode(kVK_Control), keyDown: false)
+
+        controlDown?.post(tap: .cghidEventTap)
+        arrowDown?.post(tap: .cghidEventTap)
+        arrowUp?.post(tap: .cghidEventTap)
+        controlUp?.post(tap: .cghidEventTap)
     }
 }
